feat: Add TaxRateClient class for retrieving tax rates by zip code
This commit adds a new file, `TaxRateClient.java`, which contains the implementation of a client class for retrieving tax rates by zip code. The class uses the RapidAPI service to make HTTP requests and parse JSON responses using Gson.

The `TaxRateClient` class provides a method `getTaxRateByZip(String zipCode)` that takes a zip code as input and returns a `TaxRateResponse` object containing the tax rate information for that zip code. The method sends an HTTP request to the RapidAPI endpoint with the provided zip code and retrieves the response. The response is then parsed using Gson into a `TaxRateResponse` object.

The `ShoppingCart` class has been modified to include a new instance variable `taxRate`, which stores the current tax rate. This value is set in the main method of the `ShoppingMainConsole` class by calling the new method `enterTaxZipCode(Scanner scnr)`. This method prompts the user to enter their zip code and retrieves the corresponding tax rate using an instance of the `TaxRateClient` class.

Additionally, some minor changes have been made to improve readability and maintainability:

- In the `SalesCalculator` class, there is an update to calculate percent off based on item quantity.
- In the `ShoppingCart` class, there are updates to include getter and setter methods for accessing and modifying the tax rate.
- In the `ShoppingItem` class, there are updates to include an additional parameter in methods related to calculating total price (`getTotalPrice(boolean includeTax, boolean includeShipping)`) so that shipping cost can be included or excluded based on user input.
- In the main method of