feat: Add flat shipping option to ShoppingCart
This commit adds a new field, "Flat Shipping", to the ShoppingCart class. The field is used to determine whether or not flat shipping should be applied when calculating the total price of items in the cart.

The getTotalPrice method in the ShoppingItem class has been modified to remove the includeShipping parameter and always include shipping cost if the item has shipping.

The calculateTotal method in the ShoppingCart class has been updated to add the calculated shipping cost to the total price.

The calculateShippingCost method in the ShoppingCart class has been modified to iterate through each item and sum up their individual shipping costs only if they have shipping enabled.

The toString method in the ShoppingItem class has been updated to reflect changes made in getTotalPrice calls.

In addition, there are some UI-related changes:
- The AddItemMenu and EditItemMenu classes now extend JDialog instead of JFrame.
- The AddItemMenu constructor now takes an additional parameter, mainForm, which is used for updating cart totals after adding an item.
- The EditItemMenu constructor now takes an additional parameter, mainForm, which is used for updating cart totals after editing an item.
- The ShoppingMainGUI.form file has been modified to accommodate a new label for displaying total shipping cost.